src/third_party/blink/web_tests/http/tests/devtools/elements/styles-1/add-new-rule-invalid-selector-expected.txt

Tests that adding a new rule with invalid selector works as expected.


Running: init

Running: keyframesRuleSelector
[expanded] 
element.style { ()

[expanded] 
div { (user agent stylesheet)
    display: block;

======== Inherited from html ========
[expanded] 
html { (user agent stylesheet)
    color: -internal-root-color;
